requirements.txt
tensorflow==2.9.0
numpy==1.22.4
matplotlib==3.4.2
opencv-python==4.6.0.66
pillow==9.2.0
torch==1.7.1
torchvision==0.8.2
torchoptimizer
scipy==1.9.1
scikit-image==0.19.3
scikit-learn==0.24.2
tqdm==4.66.5
argparse
glob2